refactor: main.cpp
