--- content/audio/.gitkeep
@@ -1,2 +1,0 @@
-This file is used as a mark for github to keep it (an empty folder)
-Remove this folder and replace with an readme.md (optional) when you create code in this folder
