+++ engine/core/GameLoop.cpp
@@ -1,0 +1,66 @@
+#include "GameLoop.h"
+#include "Game.h"
+#include "Window.h"
+
+#define BACKGROUND_COLOR D3DXCOLOR(200.0f/255, 200.0f/255, 255.0f/255, 1.0f)
+
+void Update(DWORD dt)
+{
+	// TODO: Handle updates for game objects / scene management
+}
+
+void Render()
+{
+	CGame* g = CGame::GetInstance();
+
+	ID3D10Device* pD3DDevice = g->GetDirect3DDevice();
+	IDXGISwapChain* pSwapChain = g->GetSwapChain();
+	ID3D10RenderTargetView* pRenderTargetView = g->GetRenderTargetView();
+	ID3DX10Sprite* spriteHandler = g->GetSpriteHandler();
+
+	pD3DDevice->ClearRenderTargetView(pRenderTargetView, BACKGROUND_COLOR);
+
+	spriteHandler->Begin(D3DX10_SPRITE_SORT_TEXTURE);
+
+	FLOAT NewBlendFactor[4] = { 0,0,0,0 };
+	pD3DDevice->OMSetBlendState(g->GetAlphaBlending(), NewBlendFactor, 0xffffffff);
+
+	// TODO: Draw logic here! e.g., sprites, tiles...
+
+	spriteHandler->End();
+	pSwapChain->Present(0, 0);
+}
+
+int Run()
+{
+	MSG msg;
+	int done = 0;
+	ULONGLONG frameStart = GetTickCount64();
+	DWORD tickPerFrame = 1000 / MAX_FRAME_RATE;
+
+	while (!done)
+	{
+		if (PeekMessage(&msg, NULL, 0, 0, PM_REMOVE))
+		{
+			if (msg.message == WM_QUIT) done = 1;
+
+			TranslateMessage(&msg);
+			DispatchMessage(&msg);
+		}
+
+		ULONGLONG now = GetTickCount64();
+		DWORD dt = (DWORD)(now - frameStart);
+
+		if (dt >= tickPerFrame)
+		{
+			frameStart = now;
+
+			Update(dt);
+			Render();
+		}
+		else
+			Sleep(tickPerFrame - dt);
+	}
+
+	return 1;
+}
