+++ engine/core/GameLoop.h
@@ -1,0 +1,6 @@
+#pragma once
+#include <windows.h>
+
+void Update(DWORD dt);
+void Render();
+int Run();
