+++ main.cpp
@@ -5,134 +5,8 @@
 #include "Engine/Utils/debug.h"
 #include "Engine/Core/Game.h"
 
-#define WINDOW_CLASS_NAME L"BaseEngineWindow"
-#define MAIN_WINDOW_TITLE L"Mario Game Base Engine"
-#define WINDOW_ICON_PATH L"mario.ico"
-
-#define BACKGROUND_COLOR D3DXCOLOR(200.0f/255, 200.0f/255, 255.0f/255, 1.0f)
-
-#define SCREEN_WIDTH 640
-#define SCREEN_HEIGHT 480
-
-LRESULT CALLBACK WinProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam)
-{
-	switch (message) {
-	case WM_DESTROY:
-		PostQuitMessage(0);
-		break;
-	default:
-		return DefWindowProc(hWnd, message, wParam, lParam);
-	}
-	return 0;
-}
-
-void Update(DWORD dt)
-{
-	// TODO: Handle updates for game objects / scene management
-}
-
-void Render()
-{
-	CGame* g = CGame::GetInstance();
-
-	ID3D10Device* pD3DDevice = g->GetDirect3DDevice();
-	IDXGISwapChain* pSwapChain = g->GetSwapChain();
-	ID3D10RenderTargetView* pRenderTargetView = g->GetRenderTargetView();
-	ID3DX10Sprite* spriteHandler = g->GetSpriteHandler();
-
-	pD3DDevice->ClearRenderTargetView(pRenderTargetView, BACKGROUND_COLOR);
-
-	spriteHandler->Begin(D3DX10_SPRITE_SORT_TEXTURE);
-
-	FLOAT NewBlendFactor[4] = { 0,0,0,0 };
-	pD3DDevice->OMSetBlendState(g->GetAlphaBlending(), NewBlendFactor, 0xffffffff);
-
-	// TODO: Draw logic here! e.g., sprites, tiles...
-
-	spriteHandler->End();
-	pSwapChain->Present(0, 0);
-}
-
-HWND CreateGameWindow(HINSTANCE hInstance, int nCmdShow, int ScreenWidth, int ScreenHeight)
-{
-	WNDCLASSEX wc;
-	wc.cbSize = sizeof(WNDCLASSEX);
-
-	wc.style = CS_HREDRAW | CS_VREDRAW;
-	wc.hInstance = hInstance;
-
-	wc.lpfnWndProc = (WNDPROC)WinProc;
-	wc.cbClsExtra = 0;
-	wc.cbWndExtra = 0;
-	wc.hIcon = (HICON)LoadImage(hInstance, WINDOW_ICON_PATH, IMAGE_ICON, 0, 0, LR_LOADFROMFILE);
-	wc.hCursor = LoadCursor(NULL, IDC_ARROW);
-	wc.hbrBackground = (HBRUSH)GetStockObject(WHITE_BRUSH);
-	wc.lpszMenuName = NULL;
-	wc.lpszClassName = WINDOW_CLASS_NAME;
-	wc.hIconSm = NULL;
-
-	RegisterClassEx(&wc);
-
-	HWND hWnd =
-		CreateWindow(
-			WINDOW_CLASS_NAME,
-			MAIN_WINDOW_TITLE,
-			WS_OVERLAPPEDWINDOW, 
-			CW_USEDEFAULT,
-			CW_USEDEFAULT,
-			ScreenWidth,
-			ScreenHeight,
-			NULL,
-			NULL,
-			hInstance,
-			NULL);
-
-	if (!hWnd) 
-	{
-		OutputDebugString(L"[ERROR] CreateWindow failed");
-		DWORD ErrCode = GetLastError();
-		return FALSE;
-	}
-
-	ShowWindow(hWnd, nCmdShow);
-	UpdateWindow(hWnd);
-
-	return hWnd;
-}
-
-int Run()
-{
-	MSG msg;
-	int done = 0;
-	ULONGLONG frameStart = GetTickCount64();
-	DWORD tickPerFrame = 1000 / MAX_FRAME_RATE;
-
-	while (!done)
-	{
-		if (PeekMessage(&msg, NULL, 0, 0, PM_REMOVE))
-		{
-			if (msg.message == WM_QUIT) done = 1;
-
-			TranslateMessage(&msg);
-			DispatchMessage(&msg);
-		}
-
-		ULONGLONG now = GetTickCount64();
-		DWORD dt = (DWORD)(now - frameStart);
-
-		if (dt >= tickPerFrame)
-		{
-			frameStart = now;
-
-			Update(dt);
-			Render();
-		}
-		else
-			Sleep(tickPerFrame - dt);	
-	}
-
-	return 1;
-}
+#include "Engine/Core/Window.h"
+#include "Engine/Core/GameLoop.h"
 
 int WINAPI WinMain(
 	_In_ HINSTANCE hInstance,
